+++ components/feature/tab-collections/src/androidTest/java/mozilla/components/feature/tab/collections/db/TabDaoTest.kt
@@ -5,17 +5,22 @@
 package mozilla.components.feature.tab.collections.db
 
 import android.content.Context
+import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.paging.PagedList
 import androidx.room.Room
 import androidx.test.core.app.ApplicationProvider
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.test.runBlockingTest
 import org.junit.After
 import org.junit.Assert.assertEquals
 import org.junit.Before
+import org.junit.Rule
 import org.junit.Test
 import java.util.UUID
 import java.util.concurrent.ExecutorService
 import java.util.concurrent.Executors
 
+@ExperimentalCoroutinesApi
 class TabDaoTest {
     private val context: Context
         get() = ApplicationProvider.getApplicationContext()
@@ -24,6 +29,9 @@ class TabDaoTest {
     private lateinit var tabCollectionDao: TabCollectionDao
     private lateinit var tabDao: TabDao
     private lateinit var executor: ExecutorService
+
+    @get:Rule
+    var instantTaskExecutorRule = InstantTaskExecutorRule()
 
     @Before
     fun setUp() {
@@ -34,7 +42,7 @@ class TabDaoTest {
     }
 
     @Test
-    fun testAddingTabsToCollection() {
+    fun testAddingTabsToCollection() = runBlockingTest {
         val collection = TabCollectionEntity(title = "Collection One", createdAt = 10).also {
             it.id = tabCollectionDao.insertTabCollection(it)
         }
@@ -45,9 +53,7 @@ class TabDaoTest {
             stateFile = UUID.randomUUID().toString(),
             tabCollectionId = collection.id!!,
             createdAt = 200
-        ).also {
-            it.id = tabDao.insertTab(it)
-        }
+        )
 
         val tab2 = TabEntity(
             title = "Tab Two",
@@ -55,9 +61,11 @@ class TabDaoTest {
             stateFile = UUID.randomUUID().toString(),
             tabCollectionId = collection.id!!,
             createdAt = 100
-        ).also {
-            it.id = tabDao.insertTab(it)
-        }
+        )
+
+        val ids = tabDao.insertTabs(listOf(tab1, tab2))
+        tab1.id = ids[0]
+        tab2.id = ids[1]
 
         val dataSource = tabCollectionDao.getTabCollectionsPaged()
             .create()
@@ -74,7 +82,7 @@ class TabDaoTest {
     }
 
     @Test
-    fun testRemovingTabFromCollection() {
+    fun testRemovingTabFromCollection() = runBlockingTest {
         val collection = TabCollectionEntity(title = "Collection One", createdAt = 10).also {
             it.id = tabCollectionDao.insertTabCollection(it)
         }
@@ -85,9 +93,7 @@ class TabDaoTest {
             stateFile = UUID.randomUUID().toString(),
             tabCollectionId = collection.id!!,
             createdAt = 200
-        ).also {
-            it.id = tabDao.insertTab(it)
-        }
+        )
 
         val tab2 = TabEntity(
             title = "Tab Two",
@@ -95,9 +101,11 @@ class TabDaoTest {
             stateFile = UUID.randomUUID().toString(),
             tabCollectionId = collection.id!!,
             createdAt = 100
-        ).also {
-            it.id = tabDao.insertTab(it)
-        }
+        )
+
+        val ids = tabDao.insertTabs(listOf(tab1, tab2))
+        tab1.id = ids[0]
+        tab2.id = ids[1]
 
         tabDao.deleteTab(tab1)