+++ components/feature/tab-collections/src/main/java/mozilla/components/feature/tab/collections/db/TabCollectionDao.kt
@@ -4,7 +4,6 @@
 
 package mozilla.components.feature.tab.collections.db
 
-import androidx.lifecycle.LiveData
 import androidx.paging.DataSource
 import androidx.room.Dao
 import androidx.room.Delete
@@ -12,6 +11,7 @@ import androidx.room.Insert
 import androidx.room.Query
 import androidx.room.Transaction
 import androidx.room.Update
+import kotlinx.coroutines.flow.Flow
 
 /**
  * Internal DAO for accessing [TabCollectionEntity] instances.
@@ -19,13 +19,13 @@ import androidx.room.Update
 @Dao
 internal interface TabCollectionDao {
     @Insert
-    fun insertTabCollection(collection: TabCollectionEntity): Long
+    suspend fun insertTabCollection(collection: TabCollectionEntity): Long
 
     @Delete
-    fun deleteTabCollection(collection: TabCollectionEntity)
+    suspend fun deleteTabCollection(collection: TabCollectionEntity)
 
     @Update
-    fun updateTabCollection(collection: TabCollectionEntity)
+    suspend fun updateTabCollection(collection: TabCollectionEntity)
 
     @Transaction
     @Query("""
@@ -43,8 +43,8 @@ internal interface TabCollectionDao {
         ORDER BY created_at DESC
         LIMIT :limit
     """)
-    fun getTabCollections(limit: Int): LiveData<List<TabCollectionWithTabs>>
+    fun getTabCollections(limit: Int): Flow<List<TabCollectionWithTabs>>
 
     @Query("SELECT COUNT(*) FROM tab_collections")
-    fun countTabCollections(): Int
+    suspend fun countTabCollections(): Int
 }